chore: Add data service and progress interceptor
This commit adds a data service and a progress interceptor to the Angular application. The data service is responsible for making HTTP requests and the progress interceptor is used to show a progress bar during HTTP requests. This will improve the user experience by providing visual feedback on the progress of the request.